requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.2
cachetools==4.2.4
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
Flask==2.3.3
frozenlist==1.4.0
google-api-core==1.34.0
google-auth==1.35.0
google-cloud-core==1.7.3
google-cloud-translate==2.0.1
googleapis-common-protos==1.60.0
grpcio==1.57.0
grpcio-status==1.48.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
libretranslatepy==2.1.1
lxml==4.9.3
MarkupSafe==2.1.3
multidict==6.0.4
openai==0.27.9
protobuf==3.20.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-dotenv==1.0.0
requests==2.31.0
rsa==4.9
six==1.16.0
tqdm==4.66.1
urllib3==2.0.4
Werkzeug==2.3.7
yarl==1.9.2
